backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
aiofiles
python-dotenv
pydantic
chromadb
sentence-transformers
google-auth
google-cloud-aiplatform
pinecone-client
pyjwt
bcrypt
sqlalchemy
asyncpg
aioredis
passlib[bcrypt]
